Add files via upload
This notebook contains the code for the competition  "house-prices-advanced-regression-techniques". It is trained with the categorical data, so I encoded all the inputs on a one-hot encoding. Since it was very difficult to the model to predict the exact price of a house, I also discretized the prices grupping them in a specified range of prices ($80.000). I created a simple feed forward model with the keras library that obtains a 46% of acuracy on training.